+++ changelog.md
@@ -134,5 +134,6 @@ Added overloads to support specifying the name of the Initial Catalog using the 
 
 ## Changes
 * The assembly now targets .NET 4.0 instead of .NET 3.5.
+* The HRESULT values from the SQL LocalDB Instance API logged in error messages are now formatted as hexadecimal values.
 * A warning is logged if the configured version of the SQL LocalDB Instance API to load instead of the latest version cannot be found.
 * Fixed ```TemporarySqlLocalDbInstance``` so it no longer throws an exception when disposed if the temporary SQL LocalDB instance could not be stopped or deleted.