+++ changelog.md
@@ -138,3 +138,5 @@ Added overloads to support specifying the name of the Initial Catalog using the 
 * A warning is logged if the configured version of the SQL LocalDB Instance API to load instead of the latest version cannot be found.
 * Fixed ```TemporarySqlLocalDbInstance``` so it no longer throws an exception when disposed if the temporary SQL LocalDB instance could not be stopped or deleted.
 * The internals now use the Windows ```LoadLibraryEx``` function instead of ```LoadLibrary```.
+
+# SqlLocalDb v1.16.0.0